android,webView: Fix scroll position on resize if at near bottom.
Webview window gets resized when keyboard pops in/out, unread notice
show/hide.

There was a issue on Android: when scroll position is near bottom and
keybaord pops out. This commit fixes it, by considering the height of
the content which is hidden below (below `clientHeight`).

Get the video showing different cases here:
https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Meeting/near/686615