android, notif, refactor: Rename `conversations` to `allConversations`
Currenlty this contains all the notitifcation it has received from the
server.

Now next we will be changing the signature of this map so that we can
segregate notification according to each account. And then individual
notification of each realm can be shown. So this map will have `Identity` as
key, which is the combination of `realm` and user account.