android, notif: Create NotificationIntentReceiver to handle ACTION_CLEAR.
Reason behind this receiver instead of existing service is: now there is
a need for sending data on clear action because it only have to clear
messages for that Identity (as each Identity have its own particular
notification i.e clear button).

I tried sending data via putExtra but am not able to receive it in
the service. So started this part of moving from service to receiver.