zulipSwitch: Derive valueControlled from props change.
Before we were only deriving `valueControlled` (which is value of the
switch) on component mount (at the time of initializing state).

But if defaultValue props gets changed, then it's value will not get
reflected. Thus when subscription add event arrives, it's value is
propagated till switch, but we aren't taking defaultValue expect at the
time of mounting.

If we want to go with the convention that defaultValue can only be used
at the first time, then we need to use value prop. Which will make
things complicated, as when switch wants to update it will receive both
defaultValue and value from parent, and it will add some logic to handle
such situations.

Actually it is not default value, rather it is original value which parent
caller wants. Basically original value before user (may) make changes.
So `originalValue` better fits here.

So use `getDerivedStateFromProps` which will derive new value of
`valueControlled` and returns an object to update the state.

> getDerivedStateFromProps is invoked right before calling the render
method, both on the initial mount and on subsequent updates.

Reference: https://reactjs.org/docs/react-component.html#static-getderivedstatefromprops

Fixes: #3106