composeBox, ios: Scroll message input text on text input if required.
Remove scroll view from input wrapper (which was added in #3595).
This fixed #3614. Also just checked it works fine on Android as well. So
basically #3595 was not a completely right solution for #3369.

Now inputs are wrapped with simple view, which is self aligned as
vertically center. ComposeMenu and SendButton are self aligned as
vertically end. And wrapper of all this three has justifyContent
as space-between, which makes them spread whole over horizontal space
available.

This removes alignItems: 'flex-end' from `composeBox`, which was pushing
input to vertically bottom, idealy it should be at vertically center.

Fixes: #3614